Show more active pokers on campaign #show page
A virtual column was created (pokes_count) on user model
to handle a more intelligent way to query 'more active users'
in the campaigns' page. A spec was added to check if the order
of appearance is correct etc. We are using a find_by_sql method
in the campaign model with an order by pokes_count to discover
more active pokers.